Item209: But why does develop require an item number for a change to a plugin?
git-svn-id: http://develop.twiki.org/svn/twiki/branches/DEVELOP/twikiplugins@6736 a00a5322-12db-0310-a70b-8735589c885e

+++ SyncContrib/unisonsync.pl
@@ -1,6 +1,12 @@
 #! perl -w
 use strict;
 use Data::Dumper;
+use File::Path;
+
+# (C) Martin Cleaver 2005
+# This may be distributed in the same terms as Perl itself.
+
+# Although protocol is a config option, the script assumes ssh
 
 my $configFile = 'unisonsync.cfg';
 require $configFile;
@@ -23,42 +29,108 @@ sub syncTWikiInstall {
 	}
 
 	print "Syncing Accounts ".join(',', @accountNames)."\n";
-#	print Dumper($accounts);
+
 
 	foreach my $accountName (@accountNames) {
 		my $account = $accounts->{$accountName};
+		
 		if (!defined $account) {
-			die "No such account '$accountName' - aborting\n";
+			$accounts->{$accountName} = {report=>''};
+			report($accounts->{$accountName}, "No such account '$accountName' - skipping\n");
+			next;
 		}
-		print "Options: for $accountName:\n".Dumper($account)."\n" if ($account->{debug} > 1);
-		print "Syncing $accountName\n" if ($account->{debug} > 0);
+
+		# Add to the structure a record of the account (user specifies it on the outside of the hash)
+		$account->{accountName}	= $accountName;
+
+		report($account, "Options: for $accountName:\n".Dumper($account)."\n") if ($account->{debug});
+		report($account, "Syncing $accountName\n") if ($account->{debug});
+		
 	    writePlinkLauncherScript($options, $account->{serverSite}, $account->{serverAccount}, $account->{clientServerPrivateKey});
 	    my @webs = @{$account->{webs}};
 	    foreach my $web (@webs) {
-	    	print "\n... Syncing data \n" if ($account->{debug} > 1);
-	    	syncFileSet($account, $options, $options->{dataDir}.'/'.$web);
-	    	print "\n... Syncing pub\n" if ($account->{debug} > 1);
-	    	syncFileSet($account, $options, $options->{pubDir}.'/'.$web);
+			syncDir($account, $options, $options->{dataDir}, $web);
+			syncDir($account, $options, $options->{pubDir}, $web);
 	    }
-		deletePlinkLauncherScript($options, $account);
+		deletePlinkLauncherScript($options, $account->{serverSite}, $account);
+	}
+
+	print "\n-----\n";
+	foreach my $accountName (@accountNames) {
+		my $account = $accounts->{$accountName};	
+		print $account->{report};
 	}
 }
 
+sub report {
+   my ($account, $text) = @_;
+   $account->{report} .= $text;
+   print $text;
+}
+
+
+sub syncDir {
+	my ($account, $options, $dir, $web) = @_;
+	report($account, "Sync report\n\nAccount:".$account->{accountName}."\nWeb: $web\nDir: $dir\n");
+
+# TODO: test you can set this to '' and have that override separation of webs.
+	unless ($account->{clientParentWeb}) {
+		$account->{clientParentWeb} = $account->{accountName};
+	}
+	
+    my $optionalClientParentSlash = optionalParentSlash($account->{clientParentWeb});
+    my $optionalServerParentSlash = optionalParentSlash($account->{serverParentWeb});
+
+	# SMELL - should not repeat this clientDirAbs
+    my $clientDir = $dir.'/'.$optionalClientParentSlash.$web;
+    my $clientDirAbs = $account->{clientRoot}.'/'.$clientDir;
+    unless (-d $clientDirAbs) {
+    	report($account, "Made directory $clientDirAbs\nWARNING: this is a new parent web - you will have to copy in the _default web content so that TWiki can show it properly\n");
+    	mkpath $clientDirAbs || report($account, "ERROR: could not make $clientDirAbs\n");
+    }
+    my $serverDir = $dir.'/'.$optionalServerParentSlash.$web;
+	report($account, "\n... Syncing $dir \n") if ($account->{debug} > 1);
+	syncFileSet($account, $options, $clientDir, $serverDir);
+}
 
 sub syncFileSet {
-   my ($account, $options, $dir) = @_;
-   my $cmd = getSyncFileSetCommand($account, $options, $dir);
-   print $cmd."\n" if ($account->{debug} > 1);
-   print `$cmd`;
+   my ($account, $options, $clientDir, $serverDir) = @_;
+
+   my $cmd = getSyncFileSetCommand($account, $options, $clientDir, $serverDir);
+   report($account, $cmd."\n") if ($account->{debug} > 1);
+   unless($account->{dryrun}) {
+	   report($account, `$cmd`);
+   } else {
+   	   report($account, "dry run, so not executing\n");
+   	   report($account, "(turn debug >= 2 to see cmd that would be executed)\n") unless ($account->{debug})
+   }
+}
+
+sub optionalParentSlash {
+   my ($optionalParent) = @_;
+    if ($optionalParent) {
+    	$optionalParent .= '/';
+    } else {
+    	$optionalParent = '';
+    }  
+    return $optionalParent;
 }
 
 sub getSyncFileSetCommand {
-    my ($account, $options, $dir) = @_;    
-    my $clientFileSet = $account->{clientRoot}.'/'.$dir;
-	my $serverSpec = $options->{protocol}.'://'.$account->{serverSite}.'/';
-    my $serverFileSet = $serverSpec.$account->{serverRoot}.'/'.$dir;
-    my @unisonClientArgs = ("-batch", "-sshcmd", $options->{plinkTempLauncherScriptFile}, $account->{unisonOptions});
-    my @cmd = ($options->{clientUnisonExecutable}, $clientFileSet, $serverFileSet, @unisonClientArgs);
+    my ($account, $options, $clientDir, $serverDir) = @_;    
+
+	my @optionalSshCmd = ();
+	my $optionalServerSpec = '';
+	if ($account->{serverSite}) {
+	   $optionalServerSpec = $options->{protocol}.'://'.$account->{serverSite}.'/';
+	   @optionalSshCmd = ("-sshcmd", $options->{plinkTempLauncherScriptFile});
+	}
+
+    my $clientFileSet = $account->{clientRoot}.'/'.$clientDir;
+    my $serverFileSet = $optionalServerSpec.$account->{serverRoot}.'/'.$serverDir;
+
+    my @unisonArgs = ("-batch", @optionalSshCmd, $account->{unisonOptions});
+    my @cmd = ($options->{clientUnisonExecutable}, $clientFileSet, $serverFileSet, @unisonArgs);
 
 	return join(' ',@cmd);
 }
@@ -79,13 +151,17 @@ n Cleaver\PuttyPrivateKey.ppk" -l mrjc -ssh unison -server -contactquietly
 
 sub writePlinkLauncherScript {
 	my ($options, $serverSite, $serverAccount, $clientServerPrivateKey) = @_;
-    open( FILE, ">$options->{plinkTempLauncherScriptFile}" );
-    print FILE plinkLauncherScriptContents($options, $serverSite, $serverAccount, $clientServerPrivateKey);
-    close(FILE);    
+	if ($serverSite) {
+	    open( FILE, ">$options->{plinkTempLauncherScriptFile}" );
+    	print FILE plinkLauncherScriptContents($options, $serverSite, $serverAccount, $clientServerPrivateKey);
+	    close(FILE);    
+	}
 }
 
 sub deletePlinkLauncherScript {
-	my ($options, $account) = @_;
-    unlink $options->{plinkTempLauncherScriptFile} unless ($account->{debug} > 2);
+	my ($options, $serverSite, $account) = @_;
+	if ($serverSite) {
+	    unlink $options->{plinkTempLauncherScriptFile} unless ($account->{debug} > 2);
+	}
 }
